Interview: uncommented invariants for creating interview by preloaded data. invariants by interview moved to separate file. KP-8030

+++ src/Tests/WB.Tests.Unit/SharedKernels/DataCollection/InterviewTests/when_creating_interview_with_preloaded_data_with_real_question.cs
@@ -73,7 +73,7 @@ namespace WB.Tests.Unit.SharedKernels.DataCollection.InterviewTests
         private static DateTime answersTime;
         private static Guid supervisorId;
         private static Guid prefilledQuestionId;
-        private static double prefilledQuestionAnswer;
+        private static decimal prefilledQuestionAnswer;
         private static Interview interview;
     }
 }
